Implement `References` and `Regex` modules
The `References` module implements functions to convert alphabetic
column references to integers and back again.

The `Regex` module implements a function to return the first instance of
a pattern in a string. This may change in a future release (e.g. a
parameter to choose between boolean, integer position or string value).